requirements.txt
django
djangorestframework
python-decouple
psycopg2-binary
drf-yasg